fix: claude mixes current source into the test-context (#46)
* fix: claude mixes current source into the test-context

The idea of gevals should be to test the functionality
of the MCP under test. When executing claude as is, the
content of the current directory is mixed up with MCP content
that is probably misleading the LLM.

* feat: show temp directory path for debugging

Show the temporary path if debug is on or there
was an execution error.